+++ src/app/dashboard/connectors/[id]/setup/SetupClient.tsx
