+++ src/app/dashboard/connectors/page.tsx
@@ -1,4 +1,4 @@
-import ConnectorsClient from "./Client";
+import ConnectorsClient from "./ConnectorsClient";
 
 export default function ConnectorsPage() {
   return <ConnectorsClient />;
