+++ src/app/dashboard/requests/[id]/setup/page.tsx
@@ -1,4 +1,4 @@
-import SetupClient from "./Client";
+import SetupClient from "./SetupClient";
 
 interface SetupPageProps {
   params: { id: string };
