+++ src/app/dashboard/requests/page.tsx
@@ -1,4 +1,4 @@
-import RequestsClient from "./Client";
+import RequestsClient from "./RequestsClient";
 
 export default function RequestsPage() {
   return <RequestsClient />;
